feat: create download link component
