+++ app/components/DownloadLink.tsx
@@ -1,0 +1,35 @@
+import styled from 'styled-components';
+import * as Sentry from '@sentry/nextjs';
+
+const StyledLink = styled.button`
+  color: ${(props) => props.theme.color.links};
+  text-decoration-line: underline;
+  word-break: break-word;
+`;
+
+const handleDownload = async (uuid, fileName) => {
+  const url = `/api/s3/download/${uuid}/${fileName}`;
+  await fetch(url)
+    .then((response) => response.json())
+    .then((response) => {
+      window.open(response, '_blank');
+    });
+};
+
+const DownloadLink = ({ uuid, fileName }) => {
+  return (
+    <StyledLink
+      data-testid="history-attachment-link"
+      onClick={(e) => {
+        e.preventDefault();
+        handleDownload(uuid, fileName).catch((error) => {
+          Sentry.captureException(error);
+        });
+      }}
+    >
+      {`${fileName}`}
+    </StyledLink>
+  );
+};
+
+export default DownloadLink;
